requirements.txt
fastapi
uvicorn[standard]
python-pptx
